+++ packages/svelte/src/lib/components/avatar/use-avatar.svelte.ts
@@ -1,3 +1,4 @@
+import { useEnvironmentContext } from '$lib/providers/environment'
 import { useLocaleContext } from '$lib/providers/locale'
 import type { Accessor, Optional } from '$lib/types'
 import { createId } from '$lib/utils/create-id'
@@ -10,9 +11,11 @@ export interface UseAvatarReturn extends Accessor<avatar.Api<PropTypes>> {}
 
 export const useAvatar = (props: UseAvatarProps = {}) => {
   const { dir } = useLocaleContext()
+  const { getRootNode } = useEnvironmentContext()
   const context = $derived({
     id: createId(),
     dir,
+    getRootNode,
     ...props,
   })