+++ packages/svelte/src/lib/providers/environment/examples/setup.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import { EnvironmentProvider } from '@ark-ui/svelte/environment'
+</script>
+
+<ifrmae title="IFrame Context">
+  <EnvironmentProvider><!-- Your App --></EnvironmentProvider>
+</ifrmae>