+++ packages/svelte/src/lib/providers/environment/index.ts
@@ -1,0 +1,9 @@
+export {
+  default as EnvironmentProvider,
+  type EnvironmentProviderProps,
+} from './environment-provider.svelte'
+export {
+  useEnvironmentContext,
+  type EnvironmentContext,
+  type RootNode,
+} from './use-environment-context'