+++ packages/svelte/CHANGELOG.md
@@ -9,6 +9,7 @@ description: All notable changes will be documented in this file.
 ## Added
 
 - Added `factory` component for `asChild` prop.
+- Added `Environment` component.
 
 ## [0.1.0] - 2024-11-27